// planning-idsdp-uaa-ui/src/app/main/modules/auth/features/model/ProfileMarksSetup.ts
export class ProfileMarksSetup {

    id: number;
    uuid: string;
    researchCategory: string;
    govEmployee: number;
    applicantAge: number;
    marksOnProfession: number;
    marksOnSubjectRelevancy: number;
    firstDivision: number;
    secDivision: number;
    thirdDivision: number;
    postGraduateResultMarksFirst: number;
    postGraduateResultMarksSec: number;
    postGraduateResultMarksThird: number;
    mphil: number;
    phd: number;
    organizationStructure: number;
    thesisGroup: number;
    nonThesis: number;
    marksIfPublishedInJournal: number;
    marksIfTrainedOnResearch: number;
    publishedJournalQtyLocal: number;
    publishedJournalQtyIntl: number;
    experienceYearsInResearchWork: number;

    //=============================================
    
    graduateCgps4: number;
    graduateCgps35To39: number;
    graduateCgpsBelow35: number;
    postGraduateCgps4: number;
    postGraduateCgps35To39: number;
    postGraduateCgpsBelow35: number;
    journalPublicationLocOneToFive: number;
    journalPublicationLocSixToTen: number;
    journalPublicationLocTenPlus: number;
    journalPublicationIntOneToFive: number;
    journalPublicationIntSixToTen: number;
    journalPublicationIntTenPlus: number;
    researchExperienceOneToSeven: number;
    researchExperienceEight: number;
    researchExperienceEightPlus: number;

}