// planning-idsdp-uaa-ui/src/app/main/modules/auth/features/model/eligibilityChecker.ts
export class EligibilityChecker {

    id: number;
    uuid: string;

    personName: string;
    stResearchCatTypeId: number;
    dateOfBirth: Date;
    graduateResult: string;
    isGovEmployee: boolean;
    postGraduateResult: string;
    isThesisGroup: boolean;
    isMPhil: boolean;
    isPhD: boolean;
    totalResearchExperience: number;
    journalPublicationLocQty: number;
    journalPublicationIntQty: number;
    isResearchTraining: boolean;
    nidNumber: string;
    nidVerificationStatus: string;

}